improve generic adapter
