+++ app/src/main/java/com/mobileforce/hometeach/ui/home/HomePageFragment.kt
@@ -96,7 +96,8 @@ class HomePageFragment : Fragment() {
 
         val onGoingAdapter = object :
             RecyclerViewAdapter<OngoingClassModelTutor>(TutorOngoingClassesAdapter.OngoingClassesDiffCallBack()) {
-            override fun getLayoutRes() = R.layout.list_item_class_ongoing_parent_dash_board
+            override fun getLayoutRes(model: OngoingClassModelTutor): Int =
+                R.layout.list_item_class_ongoing_parent_dash_board
 
             override fun getViewHolder(
                 view: View,
@@ -111,7 +112,8 @@ class HomePageFragment : Fragment() {
         }
 
         val upComingAdapter = object : RecyclerViewAdapter<UpComingClassModel>(upComingDiffUtil) {
-            override fun getLayoutRes() = R.layout.list_item_class_upcoming_parent_dash_board
+            override fun getLayoutRes(model: UpComingClassModel): Int =
+                R.layout.list_item_class_upcoming_parent_dash_board
 
             override fun getViewHolder(
                 view: View,
@@ -126,7 +128,8 @@ class HomePageFragment : Fragment() {
         }
 
         val topTutorsAdapter = object : RecyclerViewAdapter<TopTutorModel>(topTutorDiffUtil) {
-            override fun getLayoutRes() = R.layout.list_item_tutor_parent_dash_board
+            override fun getLayoutRes(model: TopTutorModel): Int =
+                R.layout.list_item_tutor_parent_dash_board
 
             override fun getViewHolder(
                 view: View,
