+++ app/src/main/java/com/mobileforce/hometeach/ui/message/ChatFragment.kt
@@ -4,12 +4,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.EditText
-import android.widget.ImageButton
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.google.firebase.firestore.FieldValue
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.mobileforce.hometeach.R
@@ -20,8 +17,6 @@ import com.mobileforce.hometeach.models.Message
 import com.mobileforce.hometeach.ui.BottomNavigationActivity
 import com.mobileforce.hometeach.utils.loadImage
 import com.mobileforce.hometeach.utils.toast
-import de.hdodenhof.circleimageview.CircleImageView
-import kotlinx.android.synthetic.main.view_message_input.*
 import org.koin.android.viewmodel.ext.android.sharedViewModel
 
 /**
@@ -70,12 +65,6 @@ class ChatFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val senderImage = view.findViewById<CircleImageView>(R.id.iv_sender_image)
-
-        val messageInput = view.findViewById<EditText>(R.id.message_input)
-
-        val sendMsg = view.findViewById<ImageButton>(R.id.send_msg)
-
         binding.vwMessageInput.addAttachment.setOnClickListener {
 //            Toast.makeText(requireContext(), "Add Attachment", Toast.LENGTH_SHORT).show()
         }
@@ -91,6 +80,8 @@ class ChatFragment : Fragment() {
         recyclerView = view.findViewById(R.id.rv_chat)
         recyclerView.setHasFixedSize(true)
         recyclerView.layoutManager = LinearLayoutManager(requireContext())
+
+
 
         viewModel.chatListItem?.let {
             binding.tvSenderName.text = it.senderName
@@ -141,103 +132,10 @@ class ChatFragment : Fragment() {
             toast("Message body is empty")
             return
         }
-
-        val messageObject = Message(message, currentUser.id)
-
-        val chatListMap =
-            hashMapOf<String, Any?>(
-                "last_message" to message,
-
-                "last_message_time" to FieldValue.serverTimestamp()
-            )
-
-
-        val lastMessageRefCurrentUser = db
-            .collection("user")
-            .document(currentUser.id)
-            .collection("connect")
-            .document(viewModel.chatListItem!!.senderId)
-
-
-        val lastMessageRefOtherUser = db
-            .collection("user")
-            .document(viewModel.chatListItem!!.senderId)
-            .collection("connect")
-            .document(currentUser.id)
-
-
-        db.collection("chat")
-            .document(viewModel.chatListItem!!.chatId)
-            .collection("message")
-            .add(messageObject)
-            .addOnCompleteListener { task ->
-
-                if (task.isSuccessful) {
-                    toast("message sent")
-                    message_input.setText("")
-
-                    db.runBatch { batch ->
-
-                        batch.update(lastMessageRefCurrentUser, chatListMap)
-
-                        batch.update(lastMessageRefOtherUser, chatListMap)
-
-                    }
-
-                }
-            }
+        viewModel.sendMessage(message).also {
+            binding.vwMessageInput.messageInput.setText("")
+        }
 
     }
-
-    // Loads dummy data to test the chat recyclerview
-//    private fun loadMessages() {
-//        val message1 = Message(
-//            "Determine whether it is a sent message or a received message",
-//            User(1, "Sola"),
-//            1001
-//        )
-//        val message2 = Message("Obtain a message from MessageList", User(2, "Bola"), 1002)
-//        val message3 =
-//            Message("Inflate the appropriate layout for the ViewHolder", User(3, "Bola"), 1003)
-//        val message4 =
-//            Message("Pass the message to the ViewHolder for binding", User(4, "Bola"), 1004)
-//        val message5 =
-//            Message("Implementing this process is standard for most types", User(5, "Bola"), 1005)
-//        val message6 = Message(
-//            "Therefore, we’ll show the code for the full adapter without further explanation",
-//            User(6, "Bola"),
-//            1006
-//        )
-//        val message7 = Message(
-//            "We’ll write these to be private inner classes of MessageListAdapter.",
-//            User(7, "Bola"),
-//            1007
-//        )
-//        val message8 = Message(
-//            "Each ViewHolder holds member views that can be bound to particular information",
-//            User(8, "Bola"),
-//            1008
-//        )
-//        val message9 = Message(
-//            "For example, TextView messageText binds to a message’s content",
-//            User(9, "Bola"),
-//            1009
-//        )
-//        val message10 = Message(
-//            "This gives view binding to the ViewHolder class rather than to onBindViewHolder",
-//            User(10, "Bola"),
-//            1010
-//        )
-//        messages.add(message1)
-//        messages.add(message2)
-//        messages.add(message3)
-//        messages.add(message4)
-//        messages.add(message5)
-//        messages.add(message6)
-//        messages.add(message7)
-//        messages.add(message8)
-//        messages.add(message9)
-//        messages.add(message10)
-//    }
 
 }
